+++ Libplanet.Tests/Store/StoreTest.cs
@@ -345,12 +345,15 @@ namespace Libplanet.Tests.Store
             Fx.Store.PutTransaction(Fx.Transaction2);
             Assert.Empty(Fx.Store.IterateStagedTransactionIds());
 
-            Fx.Store.StageTransactionIds(
-                new HashSet<TxId>()
-                {
-                    Fx.Transaction1.Id,
-                    Fx.Transaction2.Id,
-                });
+            var txIds = new HashSet<TxId>()
+            {
+                Fx.Transaction1.Id,
+                Fx.Transaction2.Id,
+            };
+
+            Dictionary<TxId, bool> toStage = txIds.ToDictionary(txId => txId, _ => true);
+
+            Fx.Store.StageTransactionIds(toStage);
             Assert.Equal(
                 new HashSet<TxId>()
                 {
@@ -370,6 +373,27 @@ namespace Libplanet.Tests.Store
                     Fx.Transaction2.Id,
                 },
                 Fx.Store.IterateStagedTransactionIds().ToHashSet());
+        }
+
+        [Fact]
+        public void IterateStagedTransactionIdsToBroadcast()
+        {
+            var toStage = new Dictionary<TxId, bool>
+            {
+                { Fx.Transaction1.Id, false },
+                { Fx.Transaction2.Id, true },
+            };
+            Fx.Store.StageTransactionIds(toStage);
+
+            Assert.Equal(
+                new HashSet<TxId> { Fx.Transaction2.Id, },
+                Fx.Store.IterateStagedTransactionIds(true).ToHashSet());
+
+            Fx.Store.UnstageTransactionIds(new HashSet<TxId> { Fx.Transaction2.Id });
+
+            Assert.Equal(
+                new HashSet<TxId>(),
+                Fx.Store.IterateStagedTransactionIds(true).ToHashSet());
         }
 
         [Fact]